+++ src/Aggregator/Domain/Repository/ArticleRepository.php
@@ -25,4 +25,6 @@ interface ArticleRepository
     public function getByLink(string $link): ?Article;
 
     public function export(?string $source, ?DateRange $date): array;
+
+    public function getLastCrawlDate(string $source, ?string $category): string;
 }